Revert "Merge branch 'group_loops2' into rect_iter3"
This reverts commit 232b3cb4500cbcc924a2376f61f37367c4c0f858, reversing
changes made to f45b8ec24aebb3093d4d3e30a66eaf7382c537d6.

+++ Sources/CvPlayer.cpp
@@ -5045,16 +5045,21 @@ bool CvPlayer::hasBusyUnit() const
 {
 	PROFILE_FUNC();
 
-	foreach_(CvSelectionGroup* pLoopSelectionGroup, groups()
-	| filtered(CvSelectionGroup::fn::isBusy()))
-	{
-		if (pLoopSelectionGroup->getNumUnits() == 0)
-		{
-			pLoopSelectionGroup->kill();
-			return false;
-		}
+	CvSelectionGroup* pLoopSelectionGroup;
+	int iLoop;
 
-		return true;
+	for(pLoopSelectionGroup = firstSelectionGroup(&iLoop); pLoopSelectionGroup; pLoopSelectionGroup = nextSelectionGroup(&iLoop))
+	{
+		if (pLoopSelectionGroup->isBusy())
+		{
+		    if (pLoopSelectionGroup->getNumUnits() == 0)
+		    {
+		        pLoopSelectionGroup->kill();
+		        return false;
+		    }
+
+			return true;
+		}
 	}
 
 	return false;
